Multiplayer is integrated
This version no longer uses xamp and instead uses NodeJS. Week 7 on
moodle gives instruction on how to use NodeJS. Right now you multiple
people can join the game bugs there are big bugs:

1) when one a browser is not selected i think the position on the server
is not updated but for some reason the sprites physics is updates
meaning it falls off screen.
2) there is no sprite for the on server players, I'm struggling to
figure out where exactly this can be set. The example uses an 'Avatar'
class which i was trying to avoid ripping off but we may need to just to
get it working. This class never states the sprite to the server in
anyway so I'm not sure how it works. It does however state the name of
the animation but that is useless without the sprite. This 'Avatar class
is inside the "Classes.js " file. I've included it for reference but the
current code doesn't utilise it.

A lot of uneccessary stuff was pulled from the example. "boot" and
"preload" were for initialising things and loading assets but we are
still doing this in the game.js file. We could move things to better
suit the examples frame work or start stripping them out.